ys/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-ninja==0.16.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pika==1.2.0
pika-stubs==0.1.3
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
